majicspace/majicspace/requirements.txt
asgiref==3.7.2
Django==5.0.1
django-admin-interface==0.28.5
django-colorfield==0.11.0
django-tinymce==3.6.1
pillow==10.2.0
psycopg2-binary==2.9.9
python-decouple==3.8
python-slugify==8.0.3
sqlparse==0.4.4
text-unidecode==1.3
typing_extensions==4.9.0
